Rename `rustc_middle::query::plumbing` to `macros`
