+++ compiler/rustc_middle/src/query/mod.rs
@@ -150,7 +150,7 @@ mod keys;
 pub mod on_disk_cache;
 pub mod system;
 #[macro_use]
-pub mod plumbing;
+pub mod macros;
 
 // Each of these queries corresponds to a function pointer field in the
 // `Providers` struct for requesting a value of that type, and a method
